Almost done
The app is functioning as desired. It gets user location and displays in depth information about the current day, gives an overview of the week and gives detailed hourly info for the whole week as well. The only thing that needs to be updated in the icon colors, from black to a darker green which matches the theme.